Remove internal only game functions
