--- src/game/mod.rs
@@ -1,3 +1,0 @@
-mod game_api;
-mod game_internal;
-pub(crate) mod game_struct;
